pull duplicated methods up
