+++ src/View/Scaffolding/Emitters/BladeSourceEmitter.php
@@ -4,8 +4,6 @@ namespace Statamic\View\Scaffolding\Emitters;
 
 use Closure;
 use Illuminate\Support\Collection;
-use Statamic\Facades\Blueprint as Blueprints;
-use Statamic\Fields\Blueprint;
 
 class BladeSourceEmitter extends AbstractSourceEmitter
 {
@@ -143,34 +141,6 @@ class BladeSourceEmitter extends AbstractSourceEmitter
     public function getVariableCounters(): array
     {
         return static::$variableCounters;
-    }
-
-    public function withCountedVariable(string|array $baseName, Closure $callback): mixed
-    {
-        if (is_array($baseName)) {
-            $varNames = [];
-            $baseNames = $baseName;
-
-            foreach ($baseNames as $name) {
-                $varNames[] = $this->getCountedVariable($name);
-            }
-
-            try {
-                return $callback(...$varNames);
-            } finally {
-                foreach (array_reverse($baseNames) as $name) {
-                    $this->releaseCountedVariable($name);
-                }
-            }
-        }
-
-        $varName = $this->getCountedVariable($baseName);
-
-        try {
-            return $callback($varName);
-        } finally {
-            $this->releaseCountedVariable($baseName);
-        }
     }
 
     public function varName(string $variable): string
@@ -541,43 +511,6 @@ BLADE_TAG;
         $this->optionalNewline();
 
         $this->content .= $this->generator()->scaffoldFields($fields);
-
-        return $this;
-    }
-
-    public function blueprint(Blueprint|string|null $blueprint, ?Closure $fallback = null): static
-    {
-        if (! $blueprint) {
-            return $this;
-        }
-
-        if (is_string($blueprint)) {
-            $blueprintHandle = $blueprint;
-            $blueprint = Blueprints::find($blueprint);
-        } else {
-            $blueprintHandle = $blueprint->handle();
-        }
-
-        if (! $blueprint) {
-            return $this;
-        }
-
-        if (in_array($blueprintHandle, static::$blueprintStack)) {
-            if ($fallback) {
-                $this->content .= $fallback($this->createInstance());
-            }
-
-            return $this;
-        }
-
-        static::$blueprintStack[] = $blueprintHandle;
-
-        try {
-            $this->optionalNewline();
-            $this->content .= $this->generator()->scaffoldBlueprint($blueprint);
-        } finally {
-            array_pop(static::$blueprintStack);
-        }
 
         return $this;
     }
